requirements.txt
flask
requests
google-genai
python-dotenv
redis
